mosquitto_api/requirements.txt
gunicorn==19.7.1
Flask==1.0.2
Flask-SQLALchemy==2.3.2
pg8000==1.12.3
paho-mqtt==1.3.1

